Some refactors, Second test for VerticalPDFViewPager
