+++ sample/src/main/AndroidManifest.xml
@@ -25,5 +25,6 @@
         <activity android:name=".ZoomablePDFActivityPhotoView" android:theme="@style/Base.Theme.AppCompat.Light.DarkActionBar"/>
         <activity android:name=".PDFWithScaleActivity" android:theme="@style/Base.Theme.AppCompat.Light.DarkActionBar"/>
         <activity android:name=".VerticalViewPagerActivity" android:theme="@style/Base.Theme.AppCompat.Light.DarkActionBar"/>
+        <activity android:name=".VerticalPDFViewPagerActivity" android:theme="@style/Base.Theme.AppCompat.Light.DarkActionBar"/>
     </application>
 </manifest>
